third_party/blink/web_tests/http/tests/cookies/multiple-cookies-expected.txt

This test checks that mulitple cookies are correctly set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check setting several cookies without clearing.
PASS cookie is 'test=foobar'.
PASS cookie is 'test2=foobar; test=foobar'.
PASS cookie is 'test2=foobar; test3=foobar; test=foobar'.
PASS successfullyParsed is true

TEST COMPLETE
